commands refactored
