+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.redis</groupId>
     <artifactId>redis-scala-netty</artifactId>
-    <version>1.0.3</version>
+    <version>1.1.0</version>
     <packaging>jar</packaging>
 
     <properties>
