+++ src/main/scala/com/fotolog/redis/RedisIO.scala
@@ -115,168 +115,15 @@ class RedisConnection(val host: String = "localhost", val port: Int = 6379) {
   }
 }
 
-private[redis] object RedisCommandEncoder {
-
-  val charset = Charset.forName("UTF-8")
-
-  val SPACE = " ".getBytes
-  val EOL = "\r\n".getBytes
-
-  val DEL = "DEL".getBytes
-  val GET = "GET".getBytes
-  val MGET = "MGET".getBytes
-  val SET = "SET".getBytes
-  val MSET = "MSET".getBytes
-  val GETSET = "GETSET".getBytes
-  val SETNX = "SETNX".getBytes
-  val MSETNX = "MSETNX".getBytes
-  val SETEX = "SETEX".getBytes
-  val INCR = "INCR".getBytes
-  val INCRBY = "INCRBY".getBytes
-  val DECR = "DECR".getBytes
-  val DECRBY = "DECRBY".getBytes
-  val APPEND = "APPEND".getBytes
-  val SUBSTR = "SUBSTR".getBytes
-  val EXPIRE = "EXPIRE".getBytes
-  val PERSIST = "PERSIST".getBytes
-  val RPUSH = "RPUSH".getBytes
-  val LPUSH = "LPUSH".getBytes
-  val LLEN = "LLEN".getBytes
-  val LRANGE = "LRANGE".getBytes
-  val LTRIM = "LTRIM".getBytes
-  val LINDEX = "LINDEX".getBytes
-  val LSET = "LSET".getBytes
-  val LREM = "LREM".getBytes
-  val LPOP = "LPOP".getBytes
-  val RPOP = "RPOP".getBytes
-  val BLPOP = "BLPOP".getBytes
-  val BRPOP = "BRPOP".getBytes
-  val RPOPLPUSH = "RPOPLPUSH".getBytes
-  val HSET = "HSET".getBytes
-  val HGET = "HGET".getBytes
-  val HMGET = "HMGET".getBytes
-  val HMSET = "HMSET".getBytes
-  val HINCRBY = "HINCRBY".getBytes
-  val HEXISTS    = "HEXISTS".getBytes
-  val HDEL = "HDEL".getBytes
-  val HLEN = "HLEN".getBytes
-  val HKEYS = "HKEYS".getBytes
-  val HVALS = "HVALS".getBytes
-  val HGETALL = "HGETALL".getBytes
-  val SADD = "SADD".getBytes
-  val SREM = "SREM".getBytes
-  val SPOP = "SPOP".getBytes
-  val SMOVE = "SMOVE".getBytes
-  val SCARD = "SCARD".getBytes
-  val SISMEMBER = "SISMEMBER".getBytes
-  val SINTER = "SINTER".getBytes
-  val SINTERSTORE = "SINTERSTORE".getBytes
-  val SUNION = "SUNION".getBytes
-  val SUNIONSTORE = "SUNIONSTORE".getBytes
-  val SDIFF = "SDIFF".getBytes
-  val SDIFFSTORE = "SDIFFSTORE".getBytes
-  val SMEMBERS = "SMEMBERS".getBytes
-  val SRANDMEMBER = "SRANDMEMBER".getBytes
-  val SORT = "SORT".getBytes
-
-  val EVAL = "EVAL".getBytes
-  val EVALSHA = "EVALSHA".getBytes
-  val SCRIPT_LOAD = "SCRIPT LOAD".getBytes
-  val SCRIPT_FLUSH = "SCRIPT FLUSH".getBytes
-  val SCRIPT_KILL = "SCRIPT KILL".getBytes
-  val SCRIPT_EXISTS = "SCRIPT EXISTS".getBytes
-
-  val PING = "PING".getBytes
-  val EXISTS = "EXISTS".getBytes
-  val TYPE = "TYPE".getBytes
-  val INFO = "INFO".getBytes
-  val FLUSHALL = "FLUSHALL".getBytes
-}
-
 @Sharable
 private[redis] class RedisCommandEncoder extends OneToOneEncoder {
   import org.jboss.netty.buffer.ChannelBuffers._
   import RedisCommandEncoder._
 
   override def encode(ctx: ChannelHandlerContext, channel: Channel, msg: AnyRef): AnyRef = {
-    //println("encode[%s]: %h -> %s".format(Thread.currentThread.getName, this, msg))
     val opFuture = msg.asInstanceOf[ResultFuture]
-    //toChannelBuffer(opFuture.cmd)
     binaryCmd(opFuture.cmd.asBin)
   }
-
-
-  //private def toChannelBuffer(cmd: Cmd): ChannelBuffer = cmd match {
-    //case Del(key) => binaryCmd(DEL, key.getBytes(charset))
-    //case Del(keys @ _*) => binaryCmd(DEL:: keys.toList.map(_.getBytes(charset)): _*)
-    //case Get(key) => binaryCmd(GET, key.getBytes(charset))
-    //case MGet(keys @ _*) => binaryCmd(MGET :: keys.toList.map(_.getBytes(charset)): _*)
-    //case Set(key, value) => binaryCmd(SET, key.getBytes(charset), value)
-    //case mset: MSet => binaryCmd(MSET :: mset.kvs.toList.map{kv => List(kv._1.getBytes(charset), kv._2)}.flatten: _*)
-    //case GetSet((key, value)) => binaryCmd(GETSET, key.getBytes(charset), value)
-    //case SetNx((key, value)) => binaryCmd(SETNX, key.getBytes(charset), value)
-    //case setNxMulti: SetNx => binaryCmd(MSETNX :: setNxMulti.kvs.toList.map{kv => List(kv._1.getBytes(charset), kv._2)}.flatten: _*)
-    //case SetEx(key, expTime, value) => binaryCmd(SETEX, key.getBytes(charset), expTime.toString.getBytes(charset), value)
-    //case Incr(key, 1) =>  binaryCmd(INCR, key.getBytes(charset))
-    //case Incr(key, delta) => binaryCmd(INCRBY, key.getBytes(charset), delta.toString.getBytes(charset))
-    //case Decr(key, 1) => binaryCmd(DECR, key.getBytes(charset))
-    //case Decr(key, delta) => binaryCmd(DECRBY, key.getBytes(charset), delta.toString.getBytes(charset))
-    //case Append((key, value)) => binaryCmd(APPEND, key.getBytes(charset), value)
-    //case Substr(key, startOffset, endOffset) => binaryCmd(SUBSTR, key.getBytes(charset), startOffset.toString.getBytes(charset), endOffset.toString.getBytes(charset))
-    //case Persist(key) => binaryCmd(PERSIST, key.getBytes(charset))
-    //case Expire(key, seconds) => binaryCmd(EXPIRE, key.getBytes(charset), seconds.toString.getBytes(charset))
-    //case Rpush((key, value)) => binaryCmd(RPUSH, key.getBytes(charset), value)
-    //case Lpush((key, value)) => binaryCmd(LPUSH, key.getBytes(charset), value)
-    //case Llen(key) => binaryCmd(LLEN, key.getBytes(charset))
-    //case Lrange(key, start, end) => binaryCmd(LRANGE, key.getBytes(charset), start.toString.getBytes(charset), end.toString.getBytes(charset))
-    //case Ltrim(key, start, end) => binaryCmd(LTRIM, key.getBytes(charset), start.toString.getBytes(charset), end.toString.getBytes(charset))
-    //case Lindex(key, idx) => binaryCmd(LINDEX, key.getBytes(charset), idx.toString.getBytes(charset))
-    //case Lset(key, idx, value) => binaryCmd(LSET, key.getBytes(charset), idx.toString.getBytes(charset), value)
-    //case Lrem(key, count, value) => binaryCmd(LREM, key.getBytes(charset), count.toString.getBytes(charset), value)
-    //case Lpop(key) => binaryCmd(LPOP, key.getBytes(charset))
-    //case Rpop(key) => binaryCmd(RPOP, key.getBytes(charset))
-    //case RpopLpush(srcKey, destKey) => binaryCmd(RPOPLPUSH, srcKey.getBytes(charset), destKey.getBytes(charset))
-    //case Hset(key, field, value) => binaryCmd(HSET, key.getBytes(charset), field.getBytes(charset), value)
-    //case Hget(key, field) => binaryCmd(HGET, key.getBytes(charset), field.getBytes(charset))
-    //case Hmget(key, fields @ _*) => binaryCmd(HMGET :: key.getBytes(charset) :: fields.toList.map{_.getBytes(charset)}: _*)
-    //case hmSet: Hmset => binaryCmd(HMSET :: hmSet.key.getBytes :: hmSet.kvs.toList.map{kv => List(kv._1.getBytes(charset), kv._2)}.flatten: _*)
-    //case Hincrby(key, field, delta) => binaryCmd(HINCRBY, key.getBytes(charset), field.getBytes(charset), delta.toString.getBytes(charset))
-    //case Hexists(key, field) => binaryCmd(HEXISTS, key.getBytes(charset), field.getBytes(charset))
-    //case Hdel(key, field) => binaryCmd(HDEL, key.getBytes(charset), field.getBytes(charset))
-    //case Hlen(key) => binaryCmd(HLEN, key.getBytes(charset))
-    //case Hkeys(key) => binaryCmd(HKEYS, key.getBytes(charset))
-    //case Hvals(key) => binaryCmd(HVALS, key.getBytes(charset))
-    //case Hgetall(key) => binaryCmd(HGETALL, key.getBytes(charset))
-    //case Sadd((key, value)) => binaryCmd(SADD, key.getBytes(charset), value)
-    //case Srem((key, value)) => binaryCmd(SREM, key.getBytes(charset), value)
-    //case Spop(key) => binaryCmd(SPOP, key.getBytes(charset))
-    //case Smove(srcKey, destKey, value) => binaryCmd(SMOVE, srcKey.getBytes(charset), destKey.getBytes(charset), value)
-    //case Scard(key) => binaryCmd(SCARD, key.getBytes(charset))
-    //case Sismember((key, value)) => binaryCmd(SISMEMBER, key.getBytes(charset), value)
-    //case Sinter(keys @ _*) => binaryCmd(SINTER :: keys.toList.map{_.getBytes(charset)}: _*)
-    //case Sinterstore(destKey, keys @ _*) => binaryCmd(SINTERSTORE :: destKey.getBytes(charset) :: keys.toList.map{_.getBytes(charset)}: _*)
-    //case Sunion(keys @ _*) => binaryCmd(SUNION:: keys.toList.map{_.getBytes(charset)}: _*)
-    //case Sunionstore(destKey, keys @ _*) => binaryCmd(SUNIONSTORE:: destKey.getBytes(charset) :: keys.toList.map{_.getBytes(charset)}: _*)
-    //case Sdiff(keys @ _*) => binaryCmd(SDIFF :: keys.toList.map{_.getBytes(charset)}: _*)
-    //case Sdiffstore(destKey, keys @ _*) => binaryCmd(SDIFFSTORE :: destKey.getBytes(charset) :: keys.toList.map{_.getBytes(charset)}: _*)
-    //case Smembers(key) => binaryCmd(SMEMBERS, key.getBytes(charset))
-    //case Srandmember(key) => binaryCmd(SRANDMEMBER, key.getBytes(charset))
-
-    //case eval: Eval => binaryCmd(EVAL :: eval.script.getBytes(charset) :: eval.kv.length.toString.getBytes :: eval.kv.toList.map{kv => List(kv._1.getBytes(charset), kv._2)}.flatten: _*)
-    //case evalsha: EvalSha => binaryCmd(EVALSHA :: evalsha.digest.getBytes(charset) :: evalsha.kv.length.toString.getBytes :: evalsha.kv.toList.map{kv => List(kv._1.getBytes(charset), kv._2)}.flatten: _*)
-
-    //case ScriptLoad(script) => binaryCmd(SCRIPT_LOAD, script.getBytes(charset))
-    //case ScriptKill() => binaryCmd(SCRIPT_KILL)
-    //case ScriptFlush() => binaryCmd(SCRIPT_FLUSH)
-    //case exists: ScriptExists => binaryCmd(SCRIPT_EXISTS, exists.script.getBytes(charset))
-
-    //case Ping() => binaryCmd(PING)
-    //case Exists(key) => binaryCmd(EXISTS, key.getBytes(charset))
-    //case Type(key) => binaryCmd(TYPE, key.getBytes(charset))
-    //case Info() => binaryCmd(INFO)
-    //case FlushAll() => binaryCmd(FLUSHALL)
-    //case x => binaryCmd(x.asBin:_*)
-  //}
 
   private def binaryCmd(cmdParts: Seq[BinVal]): ChannelBuffer = {
     val params = new Array[BinVal](3*cmdParts.length + 1)
